Completed signup functionality using passport JS
This will need to be updated when the database is setup.

Also, currently sessions are only stored in memory. This can also be
changed when the database is setup using express-sessions and
mongo-connect. At the moment, when the server crashes or is shut down,
all user sessions will be deleted and users will have to login again.