DLLT-26 completed functionality to allow admins to manually register users.
An input field is provided which allows an admin user to search through
the users database by first name or last name.

Once a user is found and selected their User ID is logged and a button
will be enabled allowing the admin to register this user.

The button sends an AJAX request to /event/:eventId/register?userId=[given user id]
and the user is registerd for the event.

This can be tested by updating the registered users tabl.