Finalising functionality allowing user to reset their password.
An email will be sent to the users email address with a unique token and
their email in a URL.

When the link is clicked the user will be able to reset their password.

The unique token is checked against the token stored in the database and
as long as the email and token match and the token has not expired, they
will be able to reset their password.